WIP: move some of the rootfs setup out of the container
Move some of the rootfs setup process out of the container init process,
into a child of the parent runc process, if it succeeds.  The parent may
have the necessary privileges to mount filesystems in a directory path
that the child, after being placed into its own user namespace, may no
longer be able to access.

This adds synchronization points for mounting the filesystem, pivoting
the root of the container's mount namespace, finalizing setup of the
root filesystem, and adds more nsenter-stle init logic to let mounting
happen in a child that's attached to all of the container's namespaces
except for its user namespace.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>